requirements.txt
Flask==2.3.3
google-generativeai==0.1.0
PyPDF2==3.0.1
python-docx==0.8.11
